:sparkles::alembic::zap: implements fit and predict and adds benchmark

+++ src/svm.rs
@@ -1,4 +1,4 @@
 pub trait SVM {
-    fn fit(&mut self, x: &Vec<Vec<f64>>, y: &Vec<f64>);
-    fn predict(&self, x: &Vec<Vec<f64>>) -> Vec<f64>;
+    fn fit(&mut self, x: &Vec<Vec<f64>>, y: &Vec<i32>);
+    fn predict(&self, x: &Vec<Vec<f64>>) -> Vec<i32>;
 }
